+++ src/whoosh/__init__.py
@@ -27,8 +27,8 @@
 
 from typing import Tuple
 
-__version__: Tuple[int, ...] = (3, 11, 4)
-__version_str__: str = "3.11.4"
+__version__: Tuple[int, ...] = (3, 11, 5)
+__version_str__: str = "3.11.5"
 
 
 def versionstring(build: bool = True, extra: bool = True) -> str: